fix: media validations that depend on associations fail
Explicitly the tag validator failed, due to missing tag association.
To avoid this in the future a new dependent for all crucial calls of `HashMediaService` or the MediaHashId factory is introduced: `HashMediaProvider`.

It will provide a static criteria factory for querying the media repository, that assign all associations that will be required. In addition, it provides a proxy to the media repostiory when searching for valid or invalid media entities in Blurhash context.