feat(command): question to select entities for scoped generation prompts when non is given
To define a generation scope when non is given, a multiple choice question promts the available folders with their names and respective assigned 'entityNames'.

The actual 'entityName' of for default folders may be hard to find for some people, so this needed to be ease up and promt the available. Beside that it does reduce the execution of the wrong scope for generation, which lead to lower run times, less wasted time and energy.