fix(message): hash generation delegation when thumbnails updated or generated
Until now, when new thumbnails were supposed to be generated or existing ones were to be updated, the Message to generate Blurhashes for the affected media was directly attached to the queue.

This could have lead to:

- an attempt to generate hashes despite the failed generation of thumbnails.
- the situation, that if the queue is processed by different handlers simultaneously, the necessary thumbnails may not have been generated yet and thus a Blurhash generation failed due to missing prerequisites.

Now, the corresponding Blurhash generation is only triggered once the thumbnail generation has been successfully processed.